Update patch set 1
Patch Set 1:

I made two changes here:

i)  I addressed a defect (https://jira.onap.org/browse/AAI-129) that was raised a couple months ago to expose the SSL protocol via the RestClient fluent API such that a specific algorithm can be set, if desired.  This parameter defaults to the namespace value of "TLS" which is influenced by the configured SSL algorithms from client-side Java JSSE properties configuration.

ii) I fixed an issue with multi-value header concatenation where the resulting output format caused an issue with the AAI X-Transaction-Id header.  Before this fix the mutli-value header was being converted to a string which then contained square brackets as part of the string value and this broke the txn-id spec in AAI.  AAI has been discarding the txnId specified by the client and using one that it generated.   Now the client is the author of the txnId that passes through the system.

Patch-set: 1