Update patch set 1
Patch Set 1:

Interesting.  Ok.  Good to know.  Originally when I tried to do the git review it failed because it said that I had to link in an issue-id.  I assumed it had to be part of the commit message.  Next time I'll add another line into the commit-msg for the issue-id.

thx.

Patch-set: 1
Label: Code-Review=0