Update patch set 2
Patch Set 2: Code-Review-1

(1 comment)

Calling the Enum method .name() does not appear to increase the code coverage measurement. A solution that does have a measurable effect is to call .valueOf() on the Enum. For example:
        RestClientMsgs.valueOf(RestClientMsgs.HTTP_RESPONSE.toString());

This artificial means of increasing the code coverage does not improve code quality. It would be better to write a functional test involving the Enum which contains an assertion (that could genuinely be falsifiable).

Patch-set: 2
Label: Code-Review=-1